# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
paths:
  /api/actors:
    get:
      summary: ''
      operationId: getApiActors
      description: ''
      parameters:
        -
          in: query
          name: per_page
          description: ''
          example: 14
          required: true
          schema:
            type: integer
            description: ''
            example: 14
        -
          in: query
          name: page
          description: ''
          example: 16
          required: true
          schema:
            type: integer
            description: ''
            example: 16
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data: []
                  links:
                    first: 'http://localhost:8000/api/actors?page=1'
                    last: 'http://localhost:8000/api/actors?page=3'
                    prev: 'http://localhost:8000/api/actors?page=15'
                    next: null
                  meta:
                    current_page: 16
                    from: null
                    last_page: 3
                    links:
                      -
                        url: 'http://localhost:8000/api/actors?page=15'
                        label: '&laquo; Previous'
                        active: false
                      -
                        url: 'http://localhost:8000/api/actors?page=1'
                        label: '1'
                        active: false
                      -
                        url: 'http://localhost:8000/api/actors?page=2'
                        label: '2'
                        active: false
                      -
                        url: 'http://localhost:8000/api/actors?page=3'
                        label: '3'
                        active: false
                      -
                        url: null
                        label: 'Next &raquo;'
                        active: false
                    path: 'http://localhost:8000/api/actors'
                    per_page: 15
                    to: null
                    total: 40
                properties:
                  data:
                    type: array
                    example: []
                  links:
                    type: object
                    properties:
                      first:
                        type: string
                        example: 'http://localhost:8000/api/actors?page=1'
                      last:
                        type: string
                        example: 'http://localhost:8000/api/actors?page=3'
                      prev:
                        type: string
                        example: 'http://localhost:8000/api/actors?page=15'
                      next:
                        type: string
                        example: null
                  meta:
                    type: object
                    properties:
                      current_page:
                        type: integer
                        example: 16
                      from:
                        type: string
                        example: null
                      last_page:
                        type: integer
                        example: 3
                      links:
                        type: array
                        example:
                          -
                            url: 'http://localhost:8000/api/actors?page=15'
                            label: '&laquo; Previous'
                            active: false
                          -
                            url: 'http://localhost:8000/api/actors?page=1'
                            label: '1'
                            active: false
                          -
                            url: 'http://localhost:8000/api/actors?page=2'
                            label: '2'
                            active: false
                          -
                            url: 'http://localhost:8000/api/actors?page=3'
                            label: '3'
                            active: false
                          -
                            url: null
                            label: 'Next &raquo;'
                            active: false
                        items:
                          type: object
                          properties:
                            url:
                              type: string
                              example: 'http://localhost:8000/api/actors?page=15'
                            label:
                              type: string
                              example: '&laquo; Previous'
                            active:
                              type: boolean
                              example: false
                      path:
                        type: string
                        example: 'http://localhost:8000/api/actors'
                      per_page:
                        type: integer
                        example: 15
                      to:
                        type: string
                        example: null
                      total:
                        type: integer
                        example: 40
      tags:
        - 'Actor management'
    post:
      summary: ''
      operationId: postApiActors
      description: ''
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
      responses: {  }
      tags:
        - 'Actor management'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Must be at least 3 characters.'
                  example: otveicaowkambgdapaqgyafqymmgwyyoahisxg
                last_name:
                  type: string
                  description: 'Must be at least 3 characters.'
                  example: rmgwckkfnpyhcrlriygdelwqksypfqmhmsqbnjscyibhrbsravgqirgfnuvntcgcnliyjiqrb
              required:
                - name
                - last_name
...
  '/api/movies/{movie_id}/actors':
    get:
      summary: ''
      operationId: getApiMoviesMovie_idActors
      description: ''
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    -
                      id: 16
                      name: 'Jammie Cormier MD'
                      lastName: Considine
                      createdAt: '2023-08-17T20:43:36.000000Z'
                    -
                      id: 15
                      name: 'Lisette Treutel'
                      lastName: Bins
                      createdAt: '2023-08-17T20:43:36.000000Z'
                    -
                      id: 15
                      name: 'Lisette Treutel'
                      lastName: Bins
                      createdAt: '2023-08-17T20:43:36.000000Z'
                    -
                      id: 16
                      name: 'Jammie Cormier MD'
                      lastName: Considine
                      createdAt: '2023-08-17T20:43:36.000000Z'
                    -
                      id: 3
                      name: 'Mckenna Quitzon IV'
                      lastName: Fritsch
                      createdAt: '2023-08-17T20:42:54.000000Z'
                  links:
                    first: 'http://localhost:8000/api/movies/1/actors?page=1'
                    last: 'http://localhost:8000/api/movies/1/actors?page=1'
                    prev: null
                    next: null
                  meta:
                    current_page: 1
                    from: 1
                    last_page: 1
                    links:
                      -
                        url: null
                        label: '&laquo; Previous'
                        active: false
                      -
                        url: 'http://localhost:8000/api/movies/1/actors?page=1'
                        label: '1'
                        active: true
                      -
                        url: null
                        label: 'Next &raquo;'
                        active: false
                    path: 'http://localhost:8000/api/movies/1/actors'
                    per_page: 15
                    to: 5
                    total: 5
                properties:
                  data:
                    type: array
                    example:
                      -
                        id: 16
                        name: 'Jammie Cormier MD'
                        lastName: Considine
                        createdAt: '2023-08-17T20:43:36.000000Z'
                      -
                        id: 15
                        name: 'Lisette Treutel'
                        lastName: Bins
                        createdAt: '2023-08-17T20:43:36.000000Z'
                      -
                        id: 15
                        name: 'Lisette Treutel'
                        lastName: Bins
                        createdAt: '2023-08-17T20:43:36.000000Z'
                      -
                        id: 16
                        name: 'Jammie Cormier MD'
                        lastName: Considine
                        createdAt: '2023-08-17T20:43:36.000000Z'
                      -
                        id: 3
                        name: 'Mckenna Quitzon IV'
                        lastName: Fritsch
                        createdAt: '2023-08-17T20:42:54.000000Z'
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 16
                        name:
                          type: string
                          example: 'Jammie Cormier MD'
                        lastName:
                          type: string
                          example: Considine
                        createdAt:
                          type: string
                          example: '2023-08-17T20:43:36.000000Z'
                  links:
                    type: object
                    properties:
                      first:
                        type: string
                        example: 'http://localhost:8000/api/movies/1/actors?page=1'
                      last:
                        type: string
                        example: 'http://localhost:8000/api/movies/1/actors?page=1'
                      prev:
                        type: string
                        example: null
                      next:
                        type: string
                        example: null
                  meta:
                    type: object
                    properties:
                      current_page:
                        type: integer
                        example: 1
                      from:
                        type: integer
                        example: 1
                      last_page:
                        type: integer
                        example: 1
                      links:
                        type: array
                        example:
                          -
                            url: null
                            label: '&laquo; Previous'
                            active: false
                          -
                            url: 'http://localhost:8000/api/movies/1/actors?page=1'
                            label: '1'
                            active: true
                          -
                            url: null
                            label: 'Next &raquo;'
                            active: false
                        items:
                          type: object
                          properties:
                            url:
                              type: string
                              example: null
                            label:
                              type: string
                              example: '&laquo; Previous'
                            active:
                              type: boolean
                              example: false
                      path:
                        type: string
                        example: 'http://localhost:8000/api/movies/1/actors'
                      per_page:
                        type: integer
                        example: 15
                      to:
                        type: integer
                        example: 5
                      total:
                        type: integer
                        example: 5
      tags:
        - 'Movie management'
    post:
      summary: ''
      operationId: postApiMoviesMovie_idActors
      description: ''
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
      responses: {  }
      tags:
        - 'Movie management'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                actors:
                  type: array
                  description: ''
                  example:
                    - quasi
                  items:
                    type: string
              required:
                - actors
    delete:
      summary: ''
      operationId: deleteApiMoviesMovie_idActors
      description: ''
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
      responses: {  }
      tags:
        - 'Movie management'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                actors:
                  type: array
                  description: ''
                  example:
                    - aut
                  items:
                    type: string
              required:
                - actors
    parameters:
      -
        in: path
        name: movie_id
        description: 'The ID of the movie.'
        example: 1
        required: true
        schema:
          type: integer
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: ''
security:
  -
    default: []